feat(eve-ng): modularize MCP server and skills
