+++ mcp-servers/eve-ng-mcp-server/mcp_init.py
@@ -1,0 +1,5 @@
+"""Shared FastMCP instance — imported by all tools_*.py modules."""
+
+from fastmcp import FastMCP
+
+mcp = FastMCP("EVE-NG MCP Server")
