+++ workspace/skills/eve-lab-topology-discovery/references/question-bank.md
@@ -1,0 +1,13 @@
+# Question bank
+
+Use these when requirements are underspecified. Ask only what is missing.
+
+1. Topology type — campus, branch/WAN, ISP/SP, data center, or hybrid?
+2. Scale — how many sites, tiers, or devices?
+3. Hierarchy / fabric style — three-tier, collapsed core, leaf-spine, hub-spoke, flat mesh?
+4. Routing / control plane — OSPF, EIGRP, BGP, IS-IS, static, or undecided?
+5. Redundancy target — single-link tolerance, single-node tolerance, or best-effort?
+6. Image / platform choices — Cisco IOL, IOS-XE CSR, vMX, vEOS, NX-OSv, VPCS, or open?
+7. Host resource limits — RAM and vCPU available for the lab?
+8. Scope — design only, design + build, or design + build + config?
+9. Existing lab — extend a `.unl`, or start from scratch?
