+++ workspace/skills/eve-lab-topology-validation/references/decision-gate.md
@@ -1,0 +1,13 @@
+# Decision gate
+
+Do not start building until all are confirmed:
+1. Topology domain classified
+2. Node list finalized
+3. Link list finalized
+4. No orphan nodes
+5. No isolated networks
+6. Redundancy level stated and accepted
+7. Forbidden adjacencies identified
+8. Image availability confirmed
+9. Host resource fit checked
+10. Scope confirmed
