+++ workspace/skills/eve-lab-topology-validation/references/output-structure.md
@@ -1,0 +1,12 @@
+# Required output structure
+
+Every completed design should include:
+1. Requirement summary
+2. Best-practice recommendations
+3. Image and platform recommendations
+4. Physical topology
+5. Logical / protocol topology
+6. Design validation check
+7. Build plan
+8. Configuration plan (if requested)
+9. Risks, open questions, unresolved items
