+++ workspace/skills/eve-ng-node-operations/references/node-types.md
@@ -1,0 +1,9 @@
+# Node type reference
+
+| node_type | Template examples | Notes |
+|---|---|---|
+| `iol` | `iol`, `iol-l2` | Fast, low RAM |
+| `qemu` | `vios`, `viosl2`, `csr1000v`, `xrv9k`, `nxosv`, `vmx`, `vsrx` | Full QEMU VMs |
+| `dynamips` | `c7200`, `c3725` | Legacy Dynamips |
+| `vpcs` | `vpcs` | Lightweight PC simulator |
+| `docker` | `alpine`, `ubuntu` | Container nodes |
